Added AutoIt script to automate the prime95.exe Found another tool (Inspect.exe from https://github.com/jvanegmond/au3_uiautomation/blob/master/Spy/Inspect.exe) that can grab for information regarding the window which reveal that most of the menu item/button can be set/select via keyboard hotkeys.
There's one issue I faced where I can focus/activate the child window (that show the worker status) so that I can copy the text to save it to a log. Anyway, I merged the main thread and this worker window into one so I don't have to bother with activating the child window.